diabetes_blood_image_project/diabetes_blood_image_project/requirements.txt
streamlit==1.39.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.3.2
opencv-python-headless==4.10.0.84
Pillow==10.2.0
matplotlib==3.9.2
joblib==1.4.2
